+++ app/models/concerns/folio/thumbnails.rb
@@ -101,6 +101,15 @@ module Folio::Thumbnails
     additional_data['animated'].presence || false
   end
 
+  def largest_thumb_key
+    keys = thumbnail_sizes.keys
+    largest_key = nil; largest_value = 0
+    keys.each do |key|
+      largest_key = key if thumbnail_sizes[key] && thumbnail_sizes[key][:height] > largest_value
+    end
+    largest_key
+  end
+
   private
 
     def reset_thumbnails