+++ test/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_04_04_130516) do
+ActiveRecord::Schema.define(version: 2019_04_18_055819) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -66,7 +66,19 @@ ActiveRecord::Schema.define(version: 2019_04_04_130516) do
     t.string "last_name"
     t.string "role"
     t.boolean "is_active", default: true
+    t.string "invitation_token"
+    t.datetime "invitation_created_at"
+    t.datetime "invitation_sent_at"
+    t.datetime "invitation_accepted_at"
+    t.integer "invitation_limit"
+    t.string "invited_by_type"
+    t.bigint "invited_by_id"
+    t.integer "invitations_count", default: 0
     t.index ["email"], name: "index_folio_accounts_on_email", unique: true
+    t.index ["invitation_token"], name: "index_folio_accounts_on_invitation_token", unique: true
+    t.index ["invitations_count"], name: "index_folio_accounts_on_invitations_count"
+    t.index ["invited_by_id"], name: "index_folio_accounts_on_invited_by_id"
+    t.index ["invited_by_type", "invited_by_id"], name: "index_folio_accounts_on_invited_by_type_and_invited_by_id"
     t.index ["reset_password_token"], name: "index_folio_accounts_on_reset_password_token", unique: true
   end